+++ packages/terra-framework-docs/src/terra-dev-site/test/worklist-data-grid/WorklistDataGridWithRowSelection.test.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import WorklistDataGrid from 'terra-worklist-data-grid';
-import gridDataJSON from './gridData.json';
+import gridDataJSON from './gridDataWithRowSelection.json';
 
 const WorklistDataGridWithRowSelection = () => {
   const rowHeaderIndex = 0;
@@ -38,6 +38,18 @@ const WorklistDataGridWithRowSelection = () => {
     clearRowSelection();
   };
 
+  const onRowSelect = (rowIdsToSelect, rowIdsToUnselect) => {
+    rows.forEach((row, index) => {
+      if (rowIdsToUnselect && rowIdsToUnselect.indexOf(row.id) >= 0) {
+        rows[index].isSelected = false;
+      }
+
+      if (rowIdsToSelect && rowIdsToSelect.indexOf(row.id) >= 0) {
+        rows[index].isSelected = true;
+      }
+    });
+  };
+
   return (
     <WorklistDataGrid
       id="default-terra-worklist-data-grid"
@@ -46,17 +58,7 @@ const WorklistDataGridWithRowSelection = () => {
       rowHeaderIndex={rowHeaderIndex}
       ariaLabel="Worklist Data Grid"
       hasSelectableRows={hasSelectableRows}
-      onRowSelect={(rowId) => {
-        const newRows = [];
-        const selectedRow = rows.find(e => e.id === rowId);
-        selectedRow.isSelected = !selectedRow.isSelected;
-        rows.forEach(element => {
-          if (element.isSelected) {
-            newRows.push(element.id);
-          }
-        });
-        setSelectedRows(determineSelectedRows(false, newRows));
-      }}
+      onRowSelect={onRowSelect}
       onRowSelectAll={() => {
         const newRows = [];
         rows.forEach(e => { e.isSelected = true; newRows.push(e.id); });
@@ -67,6 +69,9 @@ const WorklistDataGridWithRowSelection = () => {
       }}
       onDisableSelectableRows={() => {
         disableSelectableRows();
+      }}
+      onEnableRowSelection={() => {
+        setHasSelectableRows(true);
       }}
     />